+++ neo/renderer/glsl/fogShaderFP.cpp
@@ -19,7 +19,7 @@
 
 const char * const fogShaderFP = R"(
 #version 100
-precision mediump float;
+precision highp float;
   
 // In
 varying vec2 var_TexFog;            // input Fog TexCoord
@@ -28,7 +28,7 @@ varying vec2 var_TexFogEnter;       // input FogEnter TexCoord
 // Uniforms
 uniform sampler2D u_fragmentMap0;   // Fog Image
 uniform sampler2D u_fragmentMap1;   // Fog Enter Image
-uniform lowp vec4 u_fogColor;       // Fog Color
+uniform mediump vec4 u_fogColor;       // Fog Color
   
 // Out
 // gl_FragCoord                     // output Fragment color
