+++ neo/renderer/glsl/interactionPhongShaderFP.cpp
@@ -19,20 +19,20 @@
 
 const char * const interactionPhongShaderFP = R"(
 #version 100
-precision mediump float;
+precision highp float;
 
 // In
 varying vec2 var_TexDiffuse;
 varying vec2 var_TexNormal;
 varying vec2 var_TexSpecular;
 varying vec4 var_TexLight;
-varying lowp vec4 var_Color;
+varying mediump vec4 var_Color;
 varying vec3 var_L;
 varying vec3 var_V;
   
 // Uniforms
-uniform lowp vec4 u_diffuseColor;
-uniform lowp vec4 u_specularColor;
+uniform mediump vec4 u_diffuseColor;
+uniform mediump vec4 u_specularColor;
 uniform float u_specularExponent;
 uniform sampler2D u_fragmentMap0; // u_bumpTexture
 uniform sampler2D u_fragmentMap1; // u_lightFalloffTexture
