+++ neo/renderer/glsl/zfillClipShaderFP.cpp
@@ -19,7 +19,7 @@
 
 const char * const zfillClipShaderFP = R"(
 #version 100
-precision mediump float;
+precision highp float;
         
 // In
 varying vec2 var_TexDiffuse;
@@ -28,8 +28,8 @@ varying vec2 var_TexClip;
 // Uniforms
 uniform sampler2D u_fragmentMap0;
 uniform sampler2D u_fragmentMap1;
-uniform lowp float u_alphaTest;
-uniform lowp vec4 u_glColor;
+uniform mediump float u_alphaTest;
+uniform mediump vec4 u_glColor;
         
 // Out
 // gl_FragCoord
